refactoring model
