+++ conan/api/model/__init__.py
@@ -1,0 +1,3 @@
+from conan.api.model.remote import Remote
+from conan.api.model.refs import RecipeReference, PkgReference
+from conan.api.model.list import PackagesList, MultiPackagesList, ListPattern
